Allow data to be fetched from Google Places asynchronously
Store the data in a local database. Provide access to the stored data
via name of the bar.

This is not tested on the latest code because I could not get it to build
and run correctly. To fetch data (and store) the data create a new fetcher
thing and execute it. It runs asynchronously. Data can then be retrieved
from the database stuff. Right now there is just an accessor for bars by
name. We can add more as we see fit.